remedial_teaching/establishment/views.py
# -*- coding: utf-8 -*-

from .models import *
from django.http import JsonResponse
from django.utils.translation import activate, gettext_lazy as _
from rest_framework.authtoken.models import Token
from rest_framework.decorators import api_view, permission_classes, authentication_classes
from rest_framework.exceptions import AuthenticationFailed
from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework.views import APIView
from utils.utils import date_from_string
from django.conf import settings
import json

