remedial_teaching/requirements.txt
Django==3.2.12
httpie==2.6.0
httplib2==0.20.2
mysqlclient==2.2.1
pymysql==1.1.0
djangorestframework==3.14.0
requests==2.25.1
requests-toolbelt==0.9.1
django-after-response==0.2.2
django-colorfield==0.11.0
PyYAML==6.0.1
html2text==2020.1.16
django-cors-headers==4.3.1
django-googledrive-storage==1.6.0